+++ app/src/main/java/shaun/home/DropScreenAdapter.java
@@ -3,19 +3,16 @@ package shaun.home;
 import android.content.ClipData;
 import android.content.Context;
 import android.content.ContextWrapper;
-import android.content.Intent;
 import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager;
 import android.graphics.drawable.Drawable;
 import android.view.DragEvent;
 import android.view.LayoutInflater;
-import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import static android.content.Context.MODE_PRIVATE;
 
@@ -50,17 +47,6 @@ public class DropScreenAdapter extends CornerAdapter{
         }
         //set position as tag
         homeLay.setTag(position);
-//        homeLay.setOnClickListener(new View.OnClickListener() {
-//            @Override
-//            public void onClick(View v) {
-//                Context context = v.getContext();
-//                v.getTag();
-//                RelativeLayout main = ((RelativeLayout) v.getRootView().findViewById(R.id.main_activity));
-//                main.removeView(v.getRootView().findViewById(R.id.corner));
-//                Intent launchIntent = context.getPackageManager().getLaunchIntentForPackage("com.android.chrome");
-//                context.startActivity(launchIntent);
-//            }
-//        });
         homeLay.setOnDragListener(new View.OnDragListener(){
             public boolean onDrag(View v, DragEvent event) {
 
@@ -74,24 +60,13 @@ public class DropScreenAdapter extends CornerAdapter{
 //                    dragData = item.getText();
 
                         v.getContext().getSharedPreferences(name, MODE_PRIVATE).edit().putString(Integer.toString(position), (String) item.getText()).apply();
-                        // Displays a message containing the dragged data.
-                        Toast.makeText(v.getContext(), "heyyy", Toast.LENGTH_SHORT).show();
 
                         // Invalidates the view to force a redraw
                         v.invalidate();
 
-                        //kill drawer
-                        Context context = v.getContext();
-                        while (context instanceof ContextWrapper) {
-                            if (context instanceof AppDrawer) {
-                                break;
-                            }
-                            context = ((ContextWrapper)context).getBaseContext();
-                        }
-                        ((AppDrawer) context).finish();
-
                         // Returns true. DragEvent.getResult() will return true.
                         return true;
+
                     default:
                         return true;
                 }
