+++ app/src/main/java/shaun/home/corner.java
@@ -1,0 +1,6 @@
+package shaun.home;
+
+/**
+ * A placeholder fragment containing a simple view.
+ */
+public enum corner {UL,UR,LL,LR}
